[NDIN-43] SignOut/Private_route axios
